# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bloom Together API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events for the current Bloom Space
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
...
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # AVAILABILITY
  /availability:
    get:
      operationId: listAvailability
      tags: [availability]
      summary: Get availability for both members
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Availability slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Availability"
    post:
      operationId: createAvailability
      tags: [availability]
      summary: Mark availability slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityInput"
      responses:
        "201":
          description: Availability created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Availability"
...
  /availability/{id}:
    delete:
      operationId: deleteAvailability
      tags: [availability]
      summary: Remove an availability slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # VOTES
  /events/{eventId}/votes:
    get:
      operationId: listVotes
      tags: [votes]
      summary: List votes for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Votes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vote"
    post:
      operationId: createVote
      tags: [votes]
      summary: Cast a vote for an event option
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteInput"
      responses:
        "201":
          description: Vote cast
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vote"

  # MEMORIES
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        favoriteFlower:
          type: ["string", "null"]
        favoriteActivity:
          type: ["string", "null"]
        createdAt:
          type: string

    # BLOOM SPACE SCHEMAS
...
    EventInput:
      type: object
      required: [title, date, category]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        category:
          type: string
          enum: [date, hangout, study_session, work_session, trip, birthday, custom]
        isRecurring:
          type: boolean
        recurringPattern:
          type: ["string", "null"]
        reminderType:
          type: ["string", "null"]
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        category:
          type: string
          enum: [date, hangout, study_session, work_session, trip, birthday, custom]
        isRecurring:
          type: boolean
        reminderType:
          type: ["string", "null"]
    Event:
      type: object
      required: [id, title, date, category, bloomSpaceId]
      properties:
        id:
          type: integer
        bloomSpaceId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        category:
          type: string
        isRecurring:
          type: boolean
        recurringPattern:
          type: ["string", "null"]
        reminderType:
          type: ["string", "null"]
        createdAt:
          type: string

    # AVAILABILITY SCHEMAS
    AvailabilityInput:
      type: object
      required: [date, startTime, endTime]
      properties:
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
    Availability:
      type: object
      required: [id, userId, date, startTime, endTime]
      properties:
        id:
          type: integer
        userId:
          type: integer
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        user:
          $ref: "#/components/schemas/User"
...
    VoteInput:
      type: object
      required: [option]
      properties:
        option:
          type: string
    Vote:
      type: object
      required: [id, eventId, userId, option]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        userId:
          type: integer
        option:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # MEMORY SCHEMAS
...